qlearning-api/requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
python-dotenv
matplotlib
pandas
openpyxl